pyproject.toml
[tool.poetry]
name = "github-follower"
version = "0.1.0"
description = ""
authors = ["Ivan Kovtun <jeanlacroix@ukr.net>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
bs4 = "^0.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
